[Docs] Fix readthedocs for tag build (#6158)

+++ docs/source/conf.py
@@ -75,7 +75,10 @@ if READTHEDOCS_VERSION_TYPE == "tag":
     # remove the warning banner if the version is a tagged release
     header_file = os.path.join(os.path.dirname(__file__),
                                "_templates/sections/header.html")
-    os.remove(header_file)
+    # The file might be removed already if the build is triggered multiple times
+    # (readthedocs build both HTML and PDF versions separately)
+    if os.path.exists(header_file):
+        os.remove(header_file)
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
